# Add numpy benchmarks to fft and blas. 