--- a/examples/benchmarks/bench_fft.py
+++ b/examples/benchmarks/bench_fft.py
@@ -11,16 +11,46 @@
 
 
 import sys
-from time import time
-from arrayfire import (array, randu, matmul)
+from timeit import timeit
 import arrayfire as af
 
-def bench(A, iters = 100):
-    start = time()
-    for t in range(iters):
-        B = af.fft2(A)
+try:
+    import numpy as np
+except:
+    np = None
+
+
+def calc_device(n):
+    A = af.randu(n, n)
     af.sync()
-    return (time() - start) / iters
+
+    def run():
+        B = af.fft2(A)
+        af.sync()
+
+    return run
+
+
+def calc_numpy(n):
+    np.random.seed(1)
+    A = np.random.rand(n, n).astype(np.float32)
+
+    def run():
+        B = np.fft.fft2(A)
+
+    return run
+
+
+def bench(calc, iters=100, upto=13):
+    _, name = calc.__name__.split("_")
+    print("Benchmark N x N 2D fft on %s" % name)
+
+    for M in range(7, upto):
+        N = 1 << M
+        t = timeit(calc(N), number=iters) / iters
+        gflops = (10.0 * N * N * M) / (t * 1E9)
+        print("Time taken for %4d x %4d: %0.4f Gflops" % (N, N, gflops))
+
 
 if __name__ == "__main__":
 
@@ -28,13 +58,7 @@ if __name__ == "__main__":
         af.set_device(int(sys.argv[1]))
 
     af.info()
-    print("Benchmark N x N 2D fft")
 
-    for M in range(7, 13):
-        N = 1 << M
-        A = af.randu(N, N)
-        af.sync()
-
-        t = bench(A)
-        gflops = (10.0 * N * N * M) / (t * 1E9)
-        print("Time taken for %4d x %4d: %0.4f Gflops" % (N, N, gflops))
+    bench(calc_device)
+    if np:
+        bench(calc_numpy, upto=10)