--- a/examples/benchmarks/bench_fft.py
+++ b/examples/benchmarks/bench_fft.py
@@ -11,7 +11,7 @@
 
 
 import sys
-from timeit import timeit
+from time import time
 import arrayfire as af
 
 try:
@@ -24,8 +24,10 @@ def calc_arrayfire(n):
     A = af.randu(n, n)
     af.sync()
 
-    def run():
-        B = af.fft2(A)
+    def run(iters):
+        for t in range(iters):
+            B = af.fft2(A)
+
         af.sync()
 
     return run
@@ -35,8 +37,9 @@ def calc_numpy(n):
     np.random.seed(1)
     A = np.random.rand(n, n).astype(np.float32)
 
-    def run():
-        B = np.fft.fft2(A)
+    def run(iters):
+        for t in range(iters):
+            B = np.fft.fft2(A)
 
     return run
 
@@ -47,7 +50,10 @@ def bench(calc, iters=100, upto=13):
 
     for M in range(7, upto):
         N = 1 << M
-        t = timeit(calc(N), number=iters) / iters
+        run = calc(N)
+        start = time()
+        run(iters)
+        t = (time() - start) / iters
         gflops = (10.0 * N * N * M) / (t * 1E9)
         print("Time taken for %4d x %4d: %0.4f Gflops" % (N, N, gflops))