Compares the following functions

 arrayfire  |  numpy |
---|---|
 af.matmul  | np.dot     |
 af.fft2         | np.fft.fft2 |


Also switch to timeit for benchmark measurements. timeit turns off the python garbage collection during measurements, so the results may be more accurate.