@pavanky 2016-05-03
@drufat calling `af.sync()` for each run on the GPU is going to significantly slow it down. In real world code you are only ever going to call `af.sync()` occasionally. 

The proper way to benchmark it would be to call the functions inside `run` for a certain number of iterations and then call `af.sync()`