@pavanky 2016-05-03
Can you also change `bench_device` to `bench_arrayfire` because it makes more sense this way.