@drufat 2016-05-03
@pavanky My understanding is that the purpose of running multiple iterations in a benchmark is to reduce the variability in the result - the actual computation should not be modified and the comparison should still be based on the time it takes for a single run to complete. If `af.sync()` is required to do the computation for a single run, then it should be part of the benchmark as otherwise you would be unfairly favoring arrayfire by dividing some of its time over all the iteration.

Here is a comparison of the run times on my computer for the old version and the new version. 

**old bench_blas.py**
```
ArrayFire v3.3.2 (CUDA, 64-bit Linux, build default)
Platform: CUDA Toolkit 7.5, Driver: 364.19
[0] GeForce GTX 960, 2040 MB, CUDA Compute 5.2
Benchmark N x N matrix multiply
Time taken for  128 x  128: 6.0806 Gflops
Time taken for  256 x  256: 1149.6282 Gflops
Time taken for  384 x  384: 1423.8280 Gflops
Time taken for  512 x  512: 2057.0017 Gflops
Time taken for  640 x  640: 1838.6790 Gflops
Time taken for  768 x  768: 1910.5981 Gflops
Time taken for  896 x  896: 2059.2579 Gflops
Time taken for 1024 x 1024: 2405.0432 Gflops
Time taken for 1152 x 1152: 2275.2714 Gflops
Time taken for 1280 x 1280: 2331.1159 Gflops
Time taken for 1408 x 1408: 2382.6808 Gflops
Time taken for 1536 x 1536: 2511.4524 Gflops
Time taken for 1664 x 1664: 2439.3697 Gflops
Time taken for 1792 x 1792: 2468.1887 Gflops
Time taken for 1920 x 1920: 2469.8352 Gflops
Time taken for 2048 x 2048: 2563.9184 Gflops
```

**new bench_blas.py**

```
ArrayFire v3.3.2 (CUDA, 64-bit Linux, build default)
Platform: CUDA Toolkit 7.5, Driver: 364.19
[0] GeForce GTX 960, 2040 MB, CUDA Compute 5.2
Benchmark N x N matrix multiply on arrayfire
Time taken for  128 x  128: 6.0006 Gflops
Time taken for  256 x  256: 609.4753 Gflops
Time taken for  384 x  384: 947.4411 Gflops
Time taken for  512 x  512: 1546.0062 Gflops
Time taken for  640 x  640: 1759.6338 Gflops
Time taken for  768 x  768: 1812.2530 Gflops
Time taken for  896 x  896: 1928.0725 Gflops
Time taken for 1024 x 1024: 2335.1287 Gflops
Time taken for 1152 x 1152: 2193.2056 Gflops
Time taken for 1280 x 1280: 2295.2198 Gflops
Time taken for 1408 x 1408: 2338.1797 Gflops
Time taken for 1536 x 1536: 2501.7196 Gflops
Time taken for 1664 x 1664: 2403.7006 Gflops
Time taken for 1792 x 1792: 2450.4298 Gflops
Time taken for 1920 x 1920: 2455.5042 Gflops
Time taken for 2048 x 2048: 2545.7770 Gflops
Benchmark N x N matrix multiply on numpy
Time taken for  128 x  128: 3.9541 Gflops
Time taken for  256 x  256: 4.1893 Gflops
Time taken for  384 x  384: 4.2883 Gflops
Time taken for  512 x  512: 4.3267 Gflops

```