@pavanky 2016-05-03
@drufat You don't *need* to call `af.sync()`. This is only required at the *end* of a benchmark.

If you are calling it in every iteration, it is going to increase the overhead. For blas, this is a smaller overhead (about 5-10% at smaller sizes), but it is going to be higher for FFT.