@drufat 2016-05-03
Here are the benchmarks for the fft

**old bench_fft.py**
```
ArrayFire v3.3.2 (CUDA, 64-bit Linux, build default)
Platform: CUDA Toolkit 7.5, Driver: 364.19
[0] GeForce GTX 960, 2040 MB, CUDA Compute 5.2
Benchmark N x N 2D fft
Time taken for  128 x  128: 0.4803 Gflops
Time taken for  256 x  256: 79.5998 Gflops
Time taken for  512 x  512: 95.8458 Gflops
Time taken for 1024 x 1024: 119.8499 Gflops
Time taken for 2048 x 2048: 134.8493 Gflops
Time taken for 4096 x 4096: 147.1597 Gflops
```

**new bench_fft.py**
```
ArrayFire v3.3.2 (CUDA, 64-bit Linux, build default)
Platform: CUDA Toolkit 7.5, Driver: 364.19
[0] GeForce GTX 960, 2040 MB, CUDA Compute 5.2
Benchmark N x N 2D fft on arrayfire
Time taken for  128 x  128: 0.1194 Gflops
Time taken for  256 x  256: 68.5797 Gflops
Time taken for  512 x  512: 82.1921 Gflops
Time taken for 1024 x 1024: 117.4545 Gflops
Time taken for 2048 x 2048: 134.9769 Gflops
Time taken for 4096 x 4096: 147.8535 Gflops
Benchmark N x N 2D fft on numpy
Time taken for  128 x  128: 3.4478 Gflops
Time taken for  256 x  256: 3.9301 Gflops
Time taken for  512 x  512: 2.9090 Gflops

```