@pavanky 2016-05-03
@drufat It's still significantly slower at smaller sizes. Besides, always calling sync gives the wrong impression about how things should be timed on the GPU.