@drufat 2016-05-03
Alright. I switched the code back to the time module. Now `af.sync()` is only run once as per your request.